backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
requests>=2.32.0
python-dotenv>=1.1.0
